requirements.txt
fastapi[all]
pydantic
poe-api
uvicorn